requirement.txt
flask==2.1.1
newspaper3k==0.2.8
nltk==3.7
scikit-learn==1.0.1
joblib==1.1.0
gunicorn==20.1.0
